Fix "serve" command not rebuilding correctly if templates are changed
Changes to templates were not applied after a rebuild when using the "serve"
command. The cause is Twig's internal template caching that does not refresh
as the rebuild command is running in the same process that created the cache.

As no public Twig API is available to refresh the cache this fix uses a
separate process for rebuilding. Related to #321.